rm debug scaffold
