+++ src/interface/src/app/scenario-creation/step3/adv-stand-level-constraints-modal/adv-stand-level-constraints-modal.component.ts
