+++ src/interface/src/app/scenario-creation/step3/adv-stand-level-constraints/adv-stand-level-constraints.component.ts
@@ -6,6 +6,7 @@ import { DataLayersComponent } from '@app/data-layers/data-layers/data-layers.co
 import { DataLayerSelectionComponent } from '@app/plan/climate-foresight/climate-foresight-run/data-layer-selection/data-layer-selection.component';
 import { SectionComponent } from '@styleguide';
 import { AdvStandLevelConstraintsModalComponent } from '../adv-stand-level-constraints-modal/adv-stand-level-constraints-modal.component';
+import { take } from 'rxjs';
 
 @Component({
   selector: 'app-adv-stand-level-constraints',
@@ -24,9 +25,21 @@ import { AdvStandLevelConstraintsModalComponent } from '../adv-stand-level-const
 export class AdvStandLevelConstraintsComponent {
   constructor(private dialog: MatDialog) {}
 
+  // TODO: remove this:
   handleClickOpenModal() {
-    return this.dialog.open(AdvStandLevelConstraintsModalComponent, {
+    const dialogRef = this.dialog.open(AdvStandLevelConstraintsModalComponent, {
       maxWidth: '560px',
+      data: { dataLayerName: 'Data Layer Name' },
     });
+
+    dialogRef
+      .afterClosed()
+      .pipe(take(1))
+      .subscribe((confirmed) => {
+        if (confirmed) {
+          console.log('here is the result:', confirmed);
+          // store the constraint
+        }
+      });
   }
 }
