+++ src/interface/src/app/standalone/nav-bar/nav-bar.component.html
@@ -26,9 +26,7 @@
     </button>
 
     <ng-content></ng-content>
-    <!-- TODO: remove this-- it's just to test this -->
-    <a (click)="handleClickOpenModal()">TEST: open modal</a>
-
+    
     <a *ngIf="showForsysLogo" routerLink="/forsys" target="_blank">
       <img
         src="assets/png/forsys_logo.png"
