+++ src/interface/src/app/standalone/nav-bar/nav-bar.component.ts
@@ -10,7 +10,6 @@ import { PlanState } from '@plan/plan.state';
 import { BreadcrumbService } from '@services/breadcrumb.service';
 import { ButtonComponent, SectionComponent } from '@styleguide';
 import { CommonModule } from '@angular/common';
-import { AdvStandLevelConstraintsModalComponent } from '@app/scenario-creation/step3/adv-stand-level-constraints-modal/adv-stand-level-constraints-modal.component';
 
 @Component({
   standalone: true,
@@ -70,21 +69,4 @@ export class NavBarComponent implements OnInit {
     });
   }
 
-  // TODO: remove this:
-  handleClickOpenModal() {
-    const dialogRef = this.dialog.open(AdvStandLevelConstraintsModalComponent, {
-      maxWidth: '560px',
-      data: { dataLayerName: 'Data Layer Name' },
-    });
-
-    dialogRef
-      .afterClosed()
-      .pipe(take(1))
-      .subscribe((confirmed) => {
-        if (confirmed) {
-          console.log('here is the result:', confirmed);
-          // store the constraint
-        }
-      });
-  }
 }
